Tesis_Fono/requirements.txt
Flask>=2.2
Flask-Cors>=3.0
mysql-connector-python>=8.0
numpy>=1.24
scipy>=1.11
scikit-fuzzy>=0.4
librosa>=0.10
reportlab>=3.6
python-dotenv>=1.0
